+++ src/pages/statusPage/left/StatusIconAndText.tsx
