+++ src/pages/statusPage/right/areas_component/AreaText.tsx
