+++ src/pages/statusPage/right/areas_component/MovingCar.tsx
